fix: added a fix to the projects DA variation to address a bug where the order of projects outputs was not consistent (#100)
BREAKING CHANGE: In order to fix this issue, the code was refactored in a way where your code engine projects may be destroyed and recreated. If upgrading from an older version of the DA and your plan has identified your projects to be recreated, bare in mind that any apps running in those projects will also be destroyed.